requirements.txt
Flask==3.0.0
Flask-Compress==1.14
flask-session
imageio==2.33.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
openpyxl
pandas==2.0.3
pillow==10.2.0
plotly==5.18.0
PyYAML
typing_extensions==4.9.0